refactor errors; clock skew is success
